requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
mcstatus==11.0.1
asyncio==3.4.3
aiohttp==3.9.5
requests==2.31.0
colorama==0.4.6
typing-extensions==4.9.0

